Refactor motor pin names for Calliope mini3
This commit refactors the motor pin names in MicroBitIO.cpp, MicroBitIO.h and MicroBitCompat.h to match the new naming convention for Calliope mini3. The prefix "M_AIN" is changed to "M_A_IN" for Motor A IN1 and IN2 pins. The corresponding ID_PINs are also updated accordingly. Additionally, a comment is added to disable P18 pin in MicroBitIO.cpp as it is not used currently.